Kernel: implement strenghtening (minimization of environments)
Each time a judgement (env, obj, fam) is pushed to the context, we
reduce it's env to the minimum number of variables.
This relies on many unspoken lemmas, since we just take into account
the free variables of the term for the moment (not those of the type, neither
of the env). Is it correct??