Kernel: make matching on partially evaluated forms possible
That is, when encountering a defined function, first check the arguments, but forget about their transformed forms, then evaluate with the pristine arguments as given, then check the result.

Note that with this, we lose the property of congruence of evaluation, i.e. we don't have:

M --> M'
==============
C[M] --> C[M']