Kernel: Check.obj returns the completely evaluated form (back to the old behavior).
We need this for infer: it should take a sliced form (otherwise capture in the lam case).
In a sense, hiding subterms behind metavariables allows to keep them stratified and closed, thus not subject to capture if we put them behind new variables.

This deactivates matching on unevaluated expressions... for now!

+ debug messages for eval (2), and export of exception