Kernel: hack to have Test_defined pass:
It was failing because commit tests check conversion of unsliced term,
and the evaluation of defined functions is such that it can lead to additional
slicing during checking of recursive calls, but the new repository is discarded
after evaluation (only the term is kept). By ensuring that all arguments of a
defined function are already sliced, it becomes correct.
It is the case in Check, but not in Conv, so we add a call to Check in Conv just
before evaluating a defined function.

:(